Move TStateMachineDiagnosticData into own file + enable V1 on Mono.
