+++ src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/AsyncTaskMethodBuilderT.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -432,9 +431,9 @@ namespace System.Runtime.CompilerServices
             {
                 if (AsyncInstrumentation.IsSupported)
                 {
-                    methodId = TStateMachineDiagnosticData.MethodId;
-                    state = TStateMachineDiagnosticData.GetState(ref StateMachine);
-                    nextContinuation = this.GetContinuationForDiagnostics;
+                    methodId = AsyncStateMachineDiagnostics<TStateMachine>.MethodId;
+                    state = AsyncStateMachineDiagnostics<TStateMachine>.GetState(ref StateMachine);
+                    nextContinuation = this.ContinuationForDiagnostics;
                     return true;
                 }
                 else
@@ -446,78 +445,6 @@ namespace System.Runtime.CompilerServices
                 }
             }
 
-            private static class TStateMachineDiagnosticData
-            {
-#if NATIVEAOT
-                // In NativeAOT we don't have reflection to resolve the method handle and state field offset.
-                // Due to the way the state machine is constructed, we can't get a direct pointer to its MoveNext method
-                // and using the interface dispatch to locate the method at slot 0 is unreliable due to Native AOT optimizations.
-                // The state field is also not guaranteed to be at a specific offset due to auto layout and Native AOT optimizations.
-                // To support this on Native AOT we would need to precompute this information in ILC and emit a
-                // hash table keyed by state machine MethodTable. At runtime we would still need to cache
-                // this data in static fields to avoid lookup cost when walking each continuation frame.
-                // On JIT these static fields are lazy evaluated and cached on initial access, but on Native AOT
-                // they will be pre-allocated, so code should be linked out when diagnostics is not supported.
-                // Given the added complexity on Native AOT, the fact that this is only used for diagnostics,
-                // and that Native AOT currently have limited asyncv1 diagnostics support in tooling, we can
-                // postpone the support until proven needed.
-                public static ulong MethodId => 0;
-                public static int GetState(ref TStateMachine? _)
-                {
-                    return -1;
-                }
-#else
-                private static readonly ulong s_methodId = ResolveMethodId();
-                private static readonly int s_resolveStateFieldOffset = ResolveStateFieldOffset();
-
-                public static ulong MethodId => s_methodId;
-
-                [MethodImpl(MethodImplOptions.AggressiveInlining)]
-                public static int GetState(ref TStateMachine? stateMachine)
-                {
-                    if (typeof(TStateMachine).IsValueType)
-                    {
-                        // Struct: state field is inline at offset within the struct
-                        return Unsafe.As<byte, int>(ref Unsafe.AddByteOffset(ref Unsafe.As<TStateMachine?, byte>(ref stateMachine), (nint)s_resolveStateFieldOffset));
-                    }
-                    else
-                    {
-                        // Class (debug builds): StateMachine is a reference, dereference to get object data
-                        if (stateMachine != null)
-                        {
-                            return Unsafe.As<byte, int>(ref Unsafe.AddByteOffset(ref RuntimeHelpers.GetRawData(stateMachine), (nint)s_resolveStateFieldOffset));
-                        }
-                    }
-
-                    return -1;
-                }
-
-                [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2090", Justification = "State machine types are always preserved.")]
-                private static ulong ResolveMethodId()
-                {
-                    MethodInfo? methodInfo = typeof(TStateMachine).GetMethod("MoveNext", BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-                    if (methodInfo != null)
-                    {
-                        return (ulong)methodInfo.MethodHandle.Value;
-                    }
-
-                    return 0;
-                }
-
-                [UnconditionalSuppressMessage("ReflectionAnalysis", "IL2090", Justification = "State machine types are always preserved.")]
-                private static int ResolveStateFieldOffset()
-                {
-                    FieldInfo? stateField = typeof(TStateMachine).GetField("<>1__state", BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-                    if (stateField != null)
-                    {
-                        Debug.Assert(stateField is RtFieldInfo, $"Expected RtFieldInfo but got {stateField.GetType().Name}");
-                        return RuntimeFieldHandle.GetInstanceFieldOffset((RtFieldInfo)stateField);
-                    }
-
-                    return 0;
-                }
-#endif
-            }
         }
 
         /// <summary>Gets the <see cref="Task{TResult}"/> for this builder.</summary>
