+++ src/libraries/System.Private.CoreLib/src/System/Runtime/CompilerServices/ValueTaskAwaiter.cs
@@ -94,10 +94,6 @@ namespace System.Runtime.CompilerServices
             }
             else if (obj != null)
             {
-                // TODO-AsyncProfiler: Optimize by using a static delegate + original box as state instead of
-                // allocating a full AsyncTaskDispatcher (Task-derived). The IValueTaskSource.OnCompleted API
-                // already takes Action<object?> + state separately, so we can use a lightweight static callback
-                // that performs PUSH/MoveNext/POP inline without the Task overhead.
                 if (AsyncTaskDispatcherInfo.AsyncProfilerInstrumentCheckPoint)
                 {
                     box = AsyncTaskDispatcher.Create(box);
